// types/form.tsx
export interface ExpertRegistrationFormErrors {
  expert_full_name?: string;
  expert_mobile_number?: string;
  expert_email_id?: string;
  expert_teaching_subjects?: string;
}

export interface StudentRegistrationFormErrors {
  student_full_name?: string;
  student_mobile_number?: string;
  student_email_id?: string;
  student_learning_subjects?: string;
}